requirements.txt
langchain==0.1.16
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==1.23.2
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
streamlit-js-eval==0.1.7
langchain_openai==0.1.3

# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2